pyproject.toml
[project]
name = "ai-sector-research-agent"
version = "0.1.0"
description = "Standalone Deep Agents research agent for AI sector analysis"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.2.6",
    "httpx>=0.28.0",
    "langchain>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph-cli[inmem]>=0.1.55",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.1",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["research_agent"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
